+++ CHANGELOG.md
@@ -9,6 +9,8 @@
     - These reduce things like `StringCSV::new(vec!["someTxId".to_string()])` to `"someTxId".into()`
 - Added example of querying an order to /examples
 - Updated dependencies, including `h2` for a security advisory, though it doesn't affect how it's used here
+- Add `err(Debug)` tracing to all `CoreKrakenClient` methods
+- Add `ret` tracing to any small responses and all trading endpoints in `CoreKrakenClient`
 
 ### v0.0.1